requirement.txt
pandas
oracledb
sqlalchemy
python-dotenv
xlsxwriter
pymysql
python-dateutil
progiter
tqdm
psutil
openpyxl
